move copy_abacusjob to comm.py
